requirements.txt
asgiref==3.5.2
beautifulsoup4==4.12.2
Django==3.2.13
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2022.1
soupsieve==2.4.1
sqlparse==0.4.2
